requirements.txt
torch
torchvision
timm
albumentations
opencv-python
pandas
scikit-learn
imagehash
pillow
tqdm
pyyaml
gdown